# back/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API - Agenciamento de Transferência de Jogadores",
    "description": "Documentação da API do sistema de agenciamento de transferência de jogadores para clubes. Rotas protegidas exigem um token JWT: faça login em /clubes/login ou /admin/login, copie o token e clique em \"Authorize\" abaixo, informando: Bearer SEU_TOKEN",
    "version": "1.0.0"
  },
  "basePath": "/",
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Informe: Bearer SEU_TOKEN (token retornado por /clubes/login ou /admin/login)"
    }
  },
  "paths": {
...
    "/jogadores/": {
      "get": {
        "tags": [
          "Jogadores"
        ],
        "summary": "Lista jogadores (com filtros opcionais)",
        "description": "Retorna jogadores cadastrados, combinando os filtros informados (todos opcionais e combináveis com AND).",
        "parameters": [
          {
            "name": "destaque",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Quando igual a true, retorna somente os jogadores com destaque = true"
          },
          {
            "name": "nome",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "Filtra por nome, clube atual ou nacionalidade contendo o texto informado"
          },
          {
            "name": "posicaoId",
            "in": "query",
            "required": false,
            "type": "integer",
            "description": "Filtra pela posição informada"
          },
          {
            "name": "valorMaximo",
            "in": "query",
            "required": false,
            "type": "number",
            "description": "Filtra jogadores com valorPedido menor ou igual ao informado"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de jogadores retornada com sucesso."
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Jogadores"
        ],
        "summary": "Cadastra um jogador",
        "description": "Cria um novo jogador e, em seguida, consulta a IA (Gemini) para complementar automaticamente pontos fortes, pontos fracos, estilo de jogo, jogador comparável e potencial de mercado. Requer autenticação de administrador.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Rafael Andrade"
                },
                "idade": {
                  "type": "number",
                  "example": 22
                },
                "nacionalidade": {
                  "type": "string",
                  "example": "Brasil"
                },
                "clubeAtual": {
                  "type": "string",
                  "example": "Grêmio Náutico"
                },
                "pernaBoa": {
                  "type": "string",
                  "example": "DESTRA"
                },
                "valorPedido": {
                  "type": "number",
                  "example": 8500000
                },
                "foto": {
                  "type": "string",
                  "example": "https://exemplo.com/foto.jpg"
                },
                "posicaoId": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Jogador cadastrado (e enriquecido pela IA) com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autenticado."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/jogadores/{id}": {
      "get": {
        "tags": [
          "Jogadores"
        ],
        "summary": "Detalha um jogador",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Código do jogador"
          }
        ],
        "responses": {
          "200": {
            "description": "Jogador encontrado."
          },
          "404": {
            "description": "Jogador não encontrado."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Jogadores"
        ],
        "summary": "Atualiza um jogador",
        "description": "Requer autenticação de administrador. Para alternar apenas o destaque, prefira PATCH /jogadores/{id}/destaque.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Código do jogador"
          }
        ],
        "responses": {
          "200": {
            "description": "Jogador atualizado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autenticado."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Jogadores"
        ],
        "summary": "Exclui um jogador",
        "description": "Requer autenticação de administrador.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Código do jogador"
          }
        ],
        "responses": {
          "200": {
            "description": "Jogador excluído com sucesso."
          },
          "400": {
            "description": "Não foi possível excluir."
          },
          "401": {
            "description": "Não autenticado."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/jogadores/{id}/destaque": {
      "patch": {
        "tags": [
          "Jogadores"
        ],
        "summary": "Alterna o destaque de um jogador",
        "description": "Atualiza apenas o campo destaque, sem precisar reenviar o restante dos dados do jogador. Requer autenticação de administrador.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Código do jogador"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "destaque": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Destaque atualizado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}